# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Ingredient Catalogue API
  description: >
    Stores catalogue entries and the reusable tags attached to them, and supports
    searching by tag. One `Ingredient` entry covers what the monolith split into
    Food and Ingredient — they are unified; any sub-distinction (raw vs
    processed, etc.) is a tag, not a field. Hand-maintained contract; keep it in
    step with the controllers under `com.example.ingredientcatalogue.Controllers`.
  version: 0.0.1
servers:
  - url: http://localhost:8082
    description: Local

tags:
  - name: ingredients
  - name: tags
  - name: nutrition

paths:
  /api/ingredients:
    get:
      tags: [ingredients]
      summary: List ingredients, optionally filtered by name fragment and/or tag
      parameters:
        - name: name
          in: query
          description: Case-insensitive substring of the ingredient name (autocomplete).
          schema: { type: string }
          example: oat
        - name: tag
          in: query
          description: Repeatable. Tag names to filter by (case-insensitive).
          schema: { type: array, items: { type: string } }
          example: ["diet:vegan", "nut-free"]
        - name: match
          in: query
          description: '`all` (default) requires every tag; `any` requires at least one.'
          schema: { type: string, enum: [all, any], default: all }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
      responses:
        '200':
          description: A page of ingredients
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageIngredientResponse' }
    post:
      tags: [ingredients]
      summary: Create an ingredient (tags are get-or-create by name)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IngredientRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngredientResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }

  /api/ingredients/by-ids:
    get:
      tags: [ingredients]
      summary: Batch lookup by id (cross-service resolution; unknown ids omitted)
      parameters:
        - name: id
          in: query
          required: true
          description: Repeatable; at most 500.
          schema: { type: array, items: { type: integer, format: int64 } }
      responses:
        '200':
          description: The ingredients that exist, in no particular order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IngredientResponse' }
        '400': { $ref: '#/components/responses/Problem' }

  /api/ingredients/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [ingredients]
      summary: Get one ingredient
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngredientResponse' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [ingredients]
      summary: Replace an ingredient's name and tag set
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IngredientRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngredientResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [ingredients]
      summary: Delete an ingredient
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }

  /api/ingredients/{id}/tags:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [ingredients]
      summary: Attach tags (get-or-create by name) to an ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagNamesRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngredientResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }

  /api/ingredients/{id}/tags/{tagName}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
      - name: tagName
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [ingredients]
      summary: Detach one tag from an ingredient (the tag itself is kept)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngredientResponse' }
        '404': { $ref: '#/components/responses/Problem' }

  /api/ingredients/{id}/nutrition:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    put:
      tags: [ingredients]
      summary: Replace an ingredient's nutrition facts (empty body clears them)
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Nutrition' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngredientResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }

  /api/ingredients/import:
    post:
      tags: [ingredients]
      summary: Bulk-create ingredients from a CSV upload
      description: >
        `multipart/form-data`, part `file`. Header row required; columns
        (case-insensitive, any order): `name` (required), `tags`
        (`;`-separated), and any of
        `basisGrams,kcal,proteinG,carbsG,fatG,fiberG,sugarG,sodiumMg`.
        Always 200 — per-row problems are in the body.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Import report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResult' }
        '400': { $ref: '#/components/responses/Problem' }

  /api/nutrition-reference:
    get:
      tags: [nutrition]
      summary: The nutrition reference (common foods, per basisGrams grams)
      parameters:
        - name: name
          in: query
          description: Case-insensitive substring of the food name.
          schema: { type: string }
      responses:
        '200':
          description: Matching reference rows, ordered by name
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NutritionReferenceResponse' }
    post:
      tags: [nutrition]
      summary: Add a nutrition reference row
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NutritionReferenceRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NutritionReferenceResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }

  /api/tags:
    get:
      tags: [tags]
      summary: List / autocomplete tags
      parameters:
        - name: prefix
          in: query
          description: Name starts with this (autocomplete).
          schema: { type: string }
        - name: namespace
          in: query
          schema: { type: string }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
      responses:
        '200':
          description: A page of tags
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageTagResponse' }
    post:
      tags: [tags]
      summary: Create a tag explicitly (with namespace / description)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }

  /api/tags/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [tags]
      summary: Get one tag
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [tags]
      summary: Rename a tag / set its namespace or description
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [tags]
      summary: Delete a tag (detaches it from every ingredient first)
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }

  /api/tags/merge:
    post:
      tags: [tags]
      summary: Fold one or more tags into another, moving every ingredient link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MergeRequest' }
      responses:
        '200':
          description: The surviving tag
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }

components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
    Sort:
      name: sort
      in: query
      description: '`property,(asc|desc)` — e.g. `name,asc`.'
      schema: { type: string }

  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }

  schemas:
    IngredientRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 255 }
        tags:
          type: array
          items: { type: string }
          description: Tag names; created if they don't exist.
        nutrition:
          $ref: '#/components/schemas/Nutrition'
          # Optional. Omit the key to leave existing nutrition untouched on update.
    IngredientResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        tags:
          type: array
          items: { $ref: '#/components/schemas/TagResponse' }
        nutrition:
          $ref: '#/components/schemas/Nutrition'
          # Present only when the ingredient has data; the field is omitted otherwise.
    Nutrition:
      type: object
      description: Nutrition facts, per basisGrams grams (default 100). Every figure optional.
      properties:
        basisGrams: { type: number, minimum: 0, description: 'Reference amount in grams (default 100).' }
        kcal:       { type: number, minimum: 0 }
        proteinG:   { type: number, minimum: 0 }
        carbsG:     { type: number, minimum: 0 }
        fatG:       { type: number, minimum: 0 }
        fiberG:     { type: number, minimum: 0 }
        sugarG:     { type: number, minimum: 0 }
        sodiumMg:   { type: number, minimum: 0 }
    ImportResult:
      type: object
      properties:
        rows:     { type: integer, description: Data rows seen (header excluded). }
        imported: { type: integer }
        skipped:  { type: integer }
        errors:
          type: array
          items:
            type: object
            properties:
              line:    { type: integer, format: int64, description: 1-based line number in the file. }
              message: { type: string }
    NutritionReferenceResponse:
      type: object
      properties:
        id:         { type: integer, format: int64 }
        name:       { type: string }
        basisGrams: { type: number }
        kcal:       { type: number, nullable: true }
        proteinG:   { type: number, nullable: true }
        carbsG:     { type: number, nullable: true }
        fatG:       { type: number, nullable: true }
        fiberG:     { type: number, nullable: true }
        sugarG:     { type: number, nullable: true }
        sodiumMg:   { type: number, nullable: true }
    NutritionReferenceRequest:
      type: object
      required: [name]
      properties:
        name:       { type: string, maxLength: 200 }
        basisGrams: { type: number, minimum: 0, nullable: true, description: Defaults to 100 when omitted. }
        kcal:       { type: number, minimum: 0, nullable: true }
        proteinG:   { type: number, minimum: 0, nullable: true }
        carbsG:     { type: number, minimum: 0, nullable: true }
        fatG:       { type: number, minimum: 0, nullable: true }
        fiberG:     { type: number, minimum: 0, nullable: true }
        sugarG:     { type: number, minimum: 0, nullable: true }
        sodiumMg:   { type: number, minimum: 0, nullable: true }
    TagRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        namespace: { type: string, maxLength: 60, nullable: true }
        description: { type: string, maxLength: 500, nullable: true }
    TagResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, description: Normalised to trimmed lower-case. }
        namespace: { type: string, nullable: true }
        description: { type: string, nullable: true }
    TagNamesRequest:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          minItems: 1
          items: { type: string }
    MergeRequest:
      type: object
      required: [from, into]
      properties:
        from:
          type: array
          minItems: 1
          items: { type: integer, format: int64 }
        into: { type: integer, format: int64 }
    ProblemDetail:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: object
          additionalProperties: { type: string }
          description: Present on 400 validation failures — field name to message.
    # Stable page envelope (`PageResponse<T>`) — hand-controlled, not Spring
    # Data's internal `Page` serialization. Flattened (not `allOf` composed) so
    # strict validators don't treat each branch as a closed object.
    PageIngredientResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last, numberOfElements, empty]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/IngredientResponse' }
        page: { type: integer, description: Zero-based index of this page. }
        size: { type: integer, description: Requested page size. }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        numberOfElements: { type: integer, description: Items actually on this page. }
        empty: { type: boolean }
    PageTagResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last, numberOfElements, empty]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/TagResponse' }
        page: { type: integer, description: Zero-based index of this page. }
        size: { type: integer, description: Requested page size. }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        numberOfElements: { type: integer, description: Items actually on this page. }
        empty: { type: boolean }